login, user functions

+++ app/Http/Controllers/TasksController.php
@@ -14,11 +14,21 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if(\Auth::check()){
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            
+            return view('users.show', $data);
+        } else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -44,15 +54,16 @@ class TasksController extends Controller
     public function store(Request $request)
     {
         $this->validate($request, [
+            'content' => 'required|max:191',
             'status' => 'required|max:10',
         ]);
         
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
         
-        return redirect('/');
+        return redirect()->back();
     }
 
     /**
@@ -95,6 +106,7 @@ class TasksController extends Controller
     public function update(Request $request, $id)
     {
         $this->validate($request, [
+            'content' => 'required|max:191',
             'status' => 'required|max:10',
         ]);
         
@@ -102,6 +114,13 @@ class TasksController extends Controller
         $task->content = $request->content;
         $task->status = $request->status;
         $task->save();
+        /*
+        これでどうなの？
+        $task = Task::updateOrCreate(['id' => $id], [
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
+        */
         
         return redirect('/');
     }
@@ -115,8 +134,11 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
         
-        return redirect('/');
+        if(\Auth::id() === $task->user_id){
+            $task->delete();
+        }
+        
+        return redirect()->back();
     }
 }
